requirements.txt
requests
requests
pandas
lxml
bs4
aiohttp
python-dotenv
selenium
opencv-python